refactor articles functionalities
