+++ server/config/variables.js
@@ -11,3 +11,4 @@ export const twitterAppId = process.env.TWITTER_APP_ID;
 export const twitterAppSecret = process.env.TWITTER_APP_SECRET;
 export const googleClientId = process.env.GOOGLE_CLIENT_ID;
 export const googleClientSecret = process.env.GOOGLE_CLIENT_SECRET;
+export const cloudinaryUrl = process.env.CLOUDINARY_URL;
