+++ server/routes/article.routes.js
@@ -1,65 +1,49 @@
-import express, { Router } from 'express';
+import { Router } from 'express';
 import {
   ArticleController,
   CommentController,
   ClapController,
   BookmarkController,
-  SearchControllers
 } from '../controllers';
-import { Authorization } from '../middleware';
+import { Authorization, UuidValidator, ArticleValidation } from '../middleware';
+import imageUpload from '../services/cloudinary.service';
 import {
   tryCatch,
-  createArticle as createarticle,
-  imageUpload,
   checkComments,
   validateCommentType,
-  doesLikeExistInCommentForUser,
-  updateArticle,
-  checkArticleExist,
-  checkAuthor,
-  shareArticleCheck,
   verifyText,
-  UuidValidator,
-  paramsValidate,
-  SearchValidators,
   deleteImage,
-  commentIdValidator,
   checkUser,
 } from '../utils';
 
 const {
-  createComments,
-  threadedComment,
-  getCommentHistory,
-  deleteComment,
-  editComment
+  createComments, threadedComment, deleteComment, editComment
 } = CommentController;
 const { createOrRemoveBookmark } = BookmarkController;
 const {
-  createArticle,
-  editArticle,
-  getOneArticle,
-  shareArticle,
-  deleteArticle,
-  getAllTags,
-  getTopArticle
+  validCreateArticle, updateArticle, checkArticleExist,
+  checkAuthor, shareArticleCheck, paramsValidate,
+  checkQueryParams, checkSpecialChars
+} = ArticleValidation;
+const {
+  createArticle, editArticle, getOneArticle,
+  shareArticle, deleteArticle, getAllTags,
+  getTopArticle, searchArticles
 } = ArticleController;
 const { validArticleId, validCommentId, validId } = UuidValidator;
-const { getArticles } = SearchControllers;
-const { checkQueryParams, checkSpecialChars } = SearchValidators;
 const { signInAuth } = Authorization;
 
 const router = new Router();
 
 router.delete('/:slug', [signInAuth, checkArticleExist, checkAuthor, deleteImage], tryCatch(deleteArticle));
-router.get('/search', [checkQueryParams, checkSpecialChars], tryCatch(getArticles));
+router.get('/search', [checkQueryParams, checkSpecialChars], tryCatch(searchArticles));
 router.get('/topfeed', tryCatch(getTopArticle));
 
 router.get('/tags', tryCatch(getAllTags));
 
 router.get('/:slug', tryCatch(getOneArticle));
 
-router.post('/', [signInAuth, imageUpload, createarticle], tryCatch(createArticle));
+router.post('/', [signInAuth, imageUpload, validCreateArticle], tryCatch(createArticle));
 
 router.post('/:slug/comment', [signInAuth, checkComments, verifyText, validateCommentType], tryCatch(createComments));
 
@@ -67,18 +51,14 @@ router.post('/:slug/comment/:commentid/thread', [signInAuth, checkComments, vali
 
 router.get('/', [paramsValidate], tryCatch(ArticleController.getAllArticles));
 
-router.delete('/:slug/comment/:commentId', signInAuth, validCommentId, tryCatch(deleteComment));
+router.delete('/:slug/comment/:commentId', [signInAuth, validCommentId], tryCatch(deleteComment));
 
 router.post('/:articleId/claps', [signInAuth, validArticleId], tryCatch(ClapController.createClap));
 router.get('/', tryCatch(ArticleController.getAllArticles));
 
-router.post('/:articleId/bookmark', signInAuth, validArticleId, tryCatch(createOrRemoveBookmark));
+router.post('/:articleId/bookmark', [signInAuth, validArticleId], tryCatch(createOrRemoveBookmark));
 
 router.put('/:slug', [signInAuth, imageUpload, checkArticleExist, checkAuthor, updateArticle], tryCatch(editArticle));
-
-router.post('/comment/:commentId/like', [signInAuth, validCommentId, doesLikeExistInCommentForUser], tryCatch(CommentController.likeComment));
-
-router.get('/comment/:commentId/history', [commentIdValidator, signInAuth], tryCatch(getCommentHistory));
 
 router.post('/:slug/share', [signInAuth, shareArticleCheck, checkArticleExist], tryCatch(shareArticle));
